requirements.txt
git+https://github.com/mongodb/mongo-python-driver.git
azure-identity